requirements-dev.txt
Flask>=2.0
Pillow
numpy
pytest
